% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_gower.R
\name{assess_gower}
\alias{assess_gower}
\title{assess_gower}
\usage{
assess_gower(clean_data, weights = rep.int(1, p), graph = TRUE)
}
\arguments{
\item{clean_data}{a dataframe}

\item{weights}{a vector}

\item{graph}{a boolean}
}
\value{
a list containing the components clean_data, gower_dist, aggl.clust.c
}
\description{
assess_gower
}
\examples{
# Example with your dataset in "inst/"
datafile <- system.file("data_test.csv", package = "strainselect")
raw_data <- read.csv(datafile, sep = ";",header = T)
prepared_data <- prepare_input(raw_data,col_select = c(7,10,13,14,17,24),
              id=1,date=17,m=FALSE,a=TRUE)
# Apply my function
assess_gower(prepared_data, graph = T)
}
