% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_profiles.R
\name{define_profiles}
\alias{define_profiles}
\title{define_profiles}
\usage{
define_profiles(data, sep = "")
}
\arguments{
\item{data}{a data frame}

\item{sep}{Separator to use between values}
}
\value{
a data frame
}
\description{
define_profiles
}
\examples{
# Example with your dataset in "inst/"
datafile <- system.file("data_test.csv", package = "strainselect")
raw_data <- read.csv(datafile, sep = ";",header = T)
prepared_data <- prepare_input(raw_data,col_select = c(7,10,13,14,17,24),
              id=1,date=17,m=FALSE,a=TRUE)
upset_data <- prepare_upset(prepared_data)
# Apply my function
define_profiles(upset_data)
}
