% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_input.R
\name{prepare_input}
\alias{prepare_input}
\title{prepare_input}
\usage{
prepare_input(
  data,
  col_select = c(1:length(data)),
  id = 1,
  date = 0,
  m = FALSE,
  a = TRUE
)
}
\arguments{
\item{data}{a dataframe}

\item{col_select}{a vector}

\item{id}{a number}

\item{date}{a number}

\item{m}{a boolean}

\item{a}{a boolean}
}
\value{
a dataframe
}
\description{
prepare_input
}
\examples{

# Example with your dataset in "inst/"
datafile <- system.file("data_test.csv", package = "strainselect")
raw_data <- read.csv(datafile, sep = ";",header = T)
# Apply my function
prepare_input(raw_data,col_select = c(7,10,13,14,17,24),
              id=1,date=17,m=FALSE,a=TRUE)
}
