% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_profiles.R
\name{select_profiles}
\alias{select_profiles}
\title{select_profiles}
\usage{
select_profiles(profiles, prepared_data)
}
\arguments{
\item{profiles}{a data frame}

\item{prepared_data}{a data frame}
}
\value{
a data frame
}
\description{
select_profiles
}
\examples{
# Example with your dataset in "inst/"
datafile <- system.file("data_test.csv", package = "strainselect")
raw_data <- read.csv(datafile, sep = ";",header = T)
prepared_data <- prepare_input(raw_data,col_select = c(7,10,13,14,17,24),
              id=1,date=17,m=FALSE,a=TRUE)
upset_data <- prepare_upset(prepared_data)
upset_profiles <- define_profiles(upset_data)
# Apply my function
select_profiles(upset_profiles,prepared_data)
}
