% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_silhouette.R
\name{select_silhouette}
\alias{select_silhouette}
\title{select_silhouette}
\usage{
select_silhouette(
  prepared_data,
  dist,
  k,
  isdiss = inherits(gower$gower_dist, "dist"),
  hc_func = "hclust",
  hc_method = "complete"
)
}
\arguments{
\item{prepared_data}{a data frame}

\item{dist}{a distance object}

\item{k}{a integer}

\item{isdiss}{logical value specifying wether x is a dissimilarity matrix}

\item{hc_func}{the hierarchical clustering function to be used}

\item{hc_method}{he agglomeration method to be used}
}
\value{
a data frame
}
\description{
select_silhouette
}
\examples{
datafile <- system.file("data_test.csv", package = "strainselect")
raw_data <- read.csv(datafile, sep = ";",header = T)
gower <- assess_gower(prepared_data, graph = T)
prepared_data <- prepare_input(raw_data,col_select = c(7,10,13,14,17,24),
              id=1,date=17,m=FALSE,a=TRUE)

 select_silhouette(dist=gower$gower_dist,prepared_data = prepared_data, k=50, isdiss = inherits(gower$gower_dist, "dist"),hc_func = "hclust", hc_method = "complete")
}
