% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_gower.R
\name{assess_gower}
\alias{assess_gower}
\title{assess_gower}
\usage{
assess_gower(
  data,
  col_select = c(1:length(data)),
  id = 1,
  date = 0,
  m = FALSE,
  a = TRUE,
  weights = rep.int(1, p),
  graph = TRUE
)
}
\arguments{
\item{data}{a dataframe}

\item{col_select}{a vector}

\item{id}{a number}

\item{date}{a number}

\item{m}{a boolean}

\item{a}{a boolean}

\item{weights}{a vector}

\item{graph}{a boolean}
}
\value{
a list containing the components clean_data, gower_dist, aggl.clust.c
}
\description{
assess_gower
}
\examples{
# Example with your dataset in "inst/"
datafile <- system.file("data_test.csv", package = "strainselect")
raw_data <- read.csv(datafile, sep = ";",header = T)
# Apply my function
assess_gower(raw_data,col_select = c(7,10,13,14,17,24), 
                      id=1,date=17,m=FALSE,a=TRUE, graph = T)
}
